Add version to binary
Use git to download the tagged version of felix. The install make task will
use the number of commits since the most recent tag to determine the version.

This uses the same make task that the repo uses for installing. When changing
the install flags, we'll only need to change the code in one place.